refactor: moved the corresponding files to the questions/__components
