+++ src/ui/_components/molecules/index.ts
@@ -4,5 +4,4 @@ export { SelectBox } from './inputs/SelectBox'
 export { FormButton } from './buttons/FormButton'
 export { Header } from './titles/Header'
 export { SectionTitle } from './titles/SectionTitle'
-export { QuestionBox } from './cards/QuestionBox'
 export { TalkCard } from './cards/TalkCard'
